Error Headers
This commit adds the ability to set custom HTTP headers for flask-restful errors.  This is necessary for us since we use an interactive terminal and without proper CORS headers, it simply dies with a CORS error instead of showing the error code properly.

It does not come with unit tests unfortunately as I'm not quite sure how to construct them in this case.  I'd appreciate the help.  Thanks!

Signed-off-by: George Sibble <gsibble@gmail.com>